Tree-sitter logic
